- ProgressBar and close btn improvement
