+++ app/src/main/java/com/oneflow/analytics/OFSurveyActivityTop.java
@@ -91,7 +91,7 @@ public class OFSurveyActivityTop extends OFSDKBaseActivity {
 
     }
 
-    @Override
+   /* @Override
     protected void onStart() {
         super.onStart();
     
@@ -199,6 +199,6 @@ public class OFSurveyActivityTop extends OFSDKBaseActivity {
         });
         initFragment();
 
-    }
+    }*/
 
 }
