+++ app/src/main/res/anim/slide_enter_from_top.xml
@@ -20,4 +20,4 @@
 <translate xmlns:android="http://schemas.android.com/apk/res/android"
     android:duration="@android:integer/config_longAnimTime"
     android:fromYDelta="-100%p"
-    android:toYDelta="10%p" />
+    android:toYDelta="1%p" />
