+++ app/src/main/res/anim/slide_exit_upward.xml
@@ -20,5 +20,5 @@
 
 <translate xmlns:android="http://schemas.android.com/apk/res/android"
     android:duration="@android:integer/config_longAnimTime"
-    android:fromYDelta="0%p"
-    android:toYDelta="-100%p" />
+    android:fromYDelta="-100%p"
+    />
