+++ app/src/main/res/layout/survey_que_fragment.xml
@@ -117,6 +117,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@id/btn_layout"
+            android:visibility="gone"
             android:gravity="center_vertical"
             android:layout_gravity="center_horizontal"
             android:padding="10dp"
