+++ app/src/main/res/layout/survey_view.xml
@@ -26,82 +26,107 @@
         android:id="@+id/main_survey_child"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        
+
         >
 
-    <RelativeLayout
-        android:id="@+id/slider_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="5dp"
-        android:visibility="gone">
-
-        <View
-            android:id="@+id/slider"
-            android:layout_width="50dp"
-            android:layout_height="5dp"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="10dp"
-            android:background="@drawable/rounded_rectangle_small" />
-    </RelativeLayout>
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/slider_layout"
-        app:cardBackgroundColor="@color/plainwhite"
-        app:cardCornerRadius="5dp"
-        app:cardElevation="14dp"
-        app:cardUseCompatPadding="true">
-        <!--android:background="@drawable/main_rounded_rectangle_new_theme"-->
         <RelativeLayout
-            android:id="@+id/view_layout"
+            android:id="@+id/slider_layout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:padding="10dp">
+            android:padding="5dp"
+            android:visibility="gone">
 
+            <View
+                android:id="@+id/slider"
+                android:layout_width="50dp"
+                android:layout_height="5dp"
+                android:layout_centerHorizontal="true"
+                android:layout_marginTop="10dp"
+                android:background="@drawable/rounded_rectangle_small" />
+        </RelativeLayout>
 
+        <androidx.cardview.widget.CardView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/slider_layout"
+            app:cardBackgroundColor="@color/plainwhite"
+            app:cardCornerRadius="5dp"
+
+            app:cardElevation="14dp"
+            app:cardUseCompatPadding="true">
+            <!--android:background="@drawable/main_rounded_rectangle_new_theme"-->
             <RelativeLayout
-                android:id="@+id/inner_child_view"
+                android:id="@+id/view_layout"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content">
+                android:layout_height="wrap_content"
+                android:padding="10dp">
 
 
-                <!--<com.oneflow.tryskysdk.customwidgets.CustomViewPager
-                    android:id="@+id/survey_viewpager"
+                <RelativeLayout
+                    android:id="@+id/inner_child_view"
                     android:layout_width="match_parent"
+                    android:layout_height="wrap_content">
+
+
+                    <!--<com.oneflow.tryskysdk.customwidgets.CustomViewPager
+                        android:id="@+id/survey_viewpager"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:background="@android:color/white" />-->
+                    <FrameLayout
+                        android:id="@+id/fragment_view"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content" />
+
+                </RelativeLayout>
+
+                <ImageView
+                    android:id="@+id/close_btn_image_view"
+                    android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:background="@android:color/white" />-->
-                <FrameLayout
-                    android:id="@+id/fragment_view"
-                    android:layout_width="match_parent"
+                    android:layout_alignParentRight="true"
+                    android:layout_marginTop="2dp"
+                    android:src="@drawable/ic_baseline_close_24" />
+
+                <LinearLayout
+                    android:id="@+id/bottom_water_mark"
+                    android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    />
+                    android:layout_alignBottom="@id/inner_child_view"
+                    android:layout_centerHorizontal="true"
+
+                    android:background="@drawable/btn_rounded_rectangle_watermark"
+                    android:baselineAligned="true"
+                    android:gravity="center_vertical"
+                    android:padding="10dp">
+
+                    <ImageView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:src="@drawable/powered_by" />
+
+                    <com.oneflow.analytics.customwidgets.OFCustomTextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginLeft="5dp"
+                        android:text="@string/powered_by"
+                        android:textColor="@color/powored_by_txt_color" />
+                </LinearLayout>
 
             </RelativeLayout>
 
-            <ImageView
-                android:id="@+id/close_btn_image_view"
-                android:layout_width="wrap_content"
+            <ProgressBar
+                android:id="@+id/pbar"
+                style="?android:attr/progressBarStyleHorizontal"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_alignParentRight="true"
-                android:layout_marginTop="2dp"
-                android:src="@drawable/ic_baseline_close_24" />
-
-        </RelativeLayout>
-
-        <ProgressBar
-            android:id="@+id/pbar"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/view_layout"
-            android:layout_gravity="bottom"
-            android:layout_marginBottom="-6dp"
-            android:indeterminate="false"
-            android:progress="100"
-            android:scaleY="1.2" />
-        <!--android:background="@color/plainwhite"-->
-    </androidx.cardview.widget.CardView>
-</RelativeLayout>
+                android:layout_below="@id/view_layout"
+                android:layout_gravity="bottom"
+                android:layout_marginBottom="-6dp"
+                android:indeterminate="false"
+                android:progress="100"
+                android:scaleY="1.2" />
+            <!--android:background="@color/plainwhite"-->
+        </androidx.cardview.widget.CardView>
+    </RelativeLayout>
 </RelativeLayout>
