+++ app/src/main/res/layout/survey_view_full_screen.xml
@@ -90,6 +90,31 @@
                 android:layout_marginTop="2dp"
                 android:src="@drawable/ic_baseline_close_24" />
 
+
+            <LinearLayout
+                android:id="@+id/bottom_water_mark"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+
+                android:gravity="center_vertical"
+                android:layout_alignParentBottom="true"
+                android:padding="10dp"
+                android:background="@drawable/btn_rounded_rectangle_watermark"
+                android:layout_centerHorizontal="true">
+
+                <ImageView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:src="@drawable/powered_by" />
+
+                <com.oneflow.analytics.customwidgets.OFCustomTextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginLeft="5dp"
+                    android:text="@string/powered_by"
+                    android:textColor="@color/powored_by_txt_color"/>
+            </LinearLayout>
+
         </RelativeLayout>
 
         <ProgressBar
