+++ app/src/main/res/values/themes.xml
@@ -61,10 +61,10 @@
 
     </style>
     <style name="DialogTransitionFromTop" >
-        <item name="android:activityOpenEnterAnimation">@anim/slide_enter_from_top</item>
-        <item name="android:activityOpenExitAnimation">@anim/slide_enter_from_top</item>
-        <item name="android:activityCloseEnterAnimation">@anim/slide_exit_upward</item>
-        <item name="android:activityCloseExitAnimation">@anim/slide_exit_upward</item>
+        <item name="android:activityOpenEnterAnimation">@anim/fade_in_sdk</item>
+        <item name="android:activityOpenExitAnimation">@anim/fade_out_sdk</item>
+        <item name="android:activityCloseEnterAnimation">@anim/slide_enter_from_top</item>
+        <item name="android:activityCloseExitAnimation">@anim/fade_out_sdk</item>
 
     </style>
     <style name="DialogTransitionCenter" >
